+++ extensions/iconify/src/components/IconActions.tsx
@@ -1,9 +1,10 @@
 import { Action, ActionPanel, Clipboard, getPreferenceValues, Icon, showToast, Toast } from "@raycast/api";
 import { copyToClipboard, toURL } from "../utils";
-import { primaryActionEnum } from "../types";
+import { iconFormatsMap } from "../utils/format-name";
+import { primaryActionEnum, Preferences } from "../types";
 import { PropsWithChildren } from "react";
 
-const { primaryAction } = getPreferenceValues<Preferences>();
+const { primaryAction, iconNameFormat } = getPreferenceValues<Preferences>();
 
 type IconActionProps = PropsWithChildren<{
   id: string;
@@ -13,6 +14,7 @@ type IconActionProps = PropsWithChildren<{
 }>;
 
 export const IconActions = ({ id, setId, dataURIIcon, svgIcon, children }: IconActionProps) => {
+  const formattedIconName = setId ? iconFormatsMap[iconNameFormat]({ setId, iconId: id }) : id;
   const paste = <Action.Paste title="Paste SVG String" content={svgIcon} />;
   const copy = <Action.CopyToClipboard title="Copy SVG String" content={svgIcon} />;
   const pasteFile = (
@@ -48,8 +50,8 @@ export const IconActions = ({ id, setId, dataURIIcon, svgIcon, children }: IconA
       }}
     />
   );
-  const pasteName = setId ? <Action.Paste title="Paste Name" content={`${setId}:${id}`} /> : null;
-  const copyName = setId ? <Action.CopyToClipboard title="Copy Name" content={`${setId}:${id}`} /> : null;
+  const pasteName = setId ? <Action.Paste title="Paste Name" content={formattedIconName} /> : null;
+  const copyName = setId ? <Action.CopyToClipboard title="Copy Name" content={formattedIconName} /> : null;
   const copyURL = setId ? <Action.CopyToClipboard title="Copy URL" content={toURL(setId, id)} /> : null;
   const copyDataURI = <Action.CopyToClipboard title="Copy Data URI" content={dataURIIcon} />;
   return (